scripts/requirements.txt
streamlit
google-generativeai
pillow
python-dotenv
torch 
transformers 
onnxruntime 
pillow 
opencv-python 
numpy 
plotly 
pandas
